Move SyncTableJob in the validate package.
